Factor out a the balance display control to be consistent across pages
Factor out the common balance display code into its own stateless
component, and add it to the receive screen.
